Enable -Wconversion
This is mainly useful for warnings relating to converting between
pointer types and integers/booleans.

There is one important bug fix buried in here in src/Segment.cc, where a
missing parameter was causing an extra call to BackupManager::sync()
every time a segment was opened.

+++ src/HashTableBenchmark.cc
@@ -235,7 +235,8 @@ main(int argc, char **argv)
     if (numberOfKeys == 0) {
         uint64_t totalEntries = numberOfCachelines *
             TestObjectMap::entriesPerCacheLine();
-        numberOfKeys = loadFactor * totalEntries;
+        numberOfKeys = static_cast<uint64_t>(loadFactor *
+                          static_cast<double>(totalEntries));
     }
 
     hashTableBenchmark(numberOfKeys, numberOfCachelines);
